Refactor: Move context imports to a centralized directory and create new context files for MapLayerList and SourceMap
